pract7/tel_guide/conf/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
sqlparse==0.4.3
tzdata==2022.2
